requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
albucore==0.0.23
albumentations==2.0.5
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
Brotli @ file:///D:/bld/brotli-split_1666788538925/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1739515848642/work/certifi
cffi @ file:///D:/bld/cffi_1671179506518/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1735929714516/work
click @ file:///D:/bld/click_1734858834134/work
cloudpickle @ file:///home/conda/feedstock_root/build_artifacts/cloudpickle_1736947526808/work
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
comm==0.2.2
contourpy @ file:///D:/bld/contourpy_1673633923174/work
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1733332471406/work
cytoolz @ file:///D:/bld/cytoolz_1666829685081/work
dask @ file:///home/conda/feedstock_root/build_artifacts/dask-core_1742597902501/work
debugpy==1.8.14
decorator==5.2.1
decord==0.6.0
defusedxml==0.7.1
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.116.1
fastjsonschema==2.21.1
filelock==3.18.0
flatbuffers==25.2.10
fonttools @ file:///D:/bld/fonttools_1680021368967/work
fqdn==1.5.1
frozenlist==1.5.0
fsspec @ file:///home/conda/feedstock_root/build_artifacts/fsspec_1743437245292/work
h11==0.14.0
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1738578511449/work
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1733211830134/work
imagecodecs @ file:///C:/b/abs_c8joqnvdi3/croot/imagecodecs_1744617105261/work
imageio @ file:///home/conda/feedstock_root/build_artifacts/imageio_1738273805233/work
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1737420181517/work
ipykernel==6.29.5
ipython==8.35.0
isoduration==20.11.0
jax==0.5.3
jaxlib==0.5.3
jedi==0.19.2
Jinja2==3.1.6
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1733736026804/work
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kaggle==1.7.4.2
kiwisolver @ file:///D:/bld/kiwisolver_1666805768319/work
lazy_loader @ file:///home/conda/feedstock_root/build_artifacts/lazy-loader_1733636780672/work
lightning-utilities==0.14.3
locket @ file:///home/conda/feedstock_root/build_artifacts/locket_1650660393415/work
MarkupSafe==3.0.2
matplotlib @ file:///C:/b/abs_ddmsoovv14/croot/matplotlib-suite_1713336426307/work
matplotlib-inline==0.1.7
mediapipe==0.10.21
mistune==3.1.3
ml_dtypes==0.5.1
mpmath==1.3.0
multidict==6.4.3
munkres==1.1.4
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_networkx_1731521053/work
notebook==7.4.0
notebook_shim==0.2.4
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.10.0
opencv-python-headless==4.10.0
opt_einsum==3.4.0
overrides==7.7.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1733203243479/work
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
partd @ file:///home/conda/feedstock_root/build_artifacts/partd_1715026491486/work
pillow @ file:///C:/b/abs_68t826txdy/croot/pillow_1744613085333/work
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1742485085/work
pooch @ file:///home/conda/feedstock_root/build_artifacts/pooch_1733421311631/work
prometheus_client==0.21.1
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==4.25.6
psutil==7.0.0
pure_eval==0.2.3
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1652235407899/work
PyQt5==5.15.4
PyQt5-sip @ file:///D:/bld/pyqt-split_1652284087426/work/pyqt_sip
PySocks @ file:///D:/bld/pysocks_1733217287171/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1733215673016/work
python-dotenv==1.1.1
python-json-logger==3.3.0
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
PyWavelets @ file:///D:/bld/pywavelets_1673082559203/work
pywin32==310
pywinpty==2.0.15
PyYAML @ file:///D:/bld/pyyaml_1666772550103/work
pyzmq==26.4.0
referencing==0.36.2
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1733217035951/work
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.24.0
scikit-image @ file:///D:/bld/scikit-image_1681313283062/work
scikit-learn @ file:///D:/bld/scikit-learn_1679676108524/work
scipy==1.15.2
seaborn==0.13.2
Send2Trash==1.8.3
sentencepiece==0.2.0
simsimd==6.2.1
sip @ file:///D:/bld/sip_1646101407041/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1733380938961/work
sniffio==1.3.1
sounddevice==0.5.1
soupsieve==2.6
stack-data==0.6.3
starlette==0.47.2
stringzilla==3.12.4
sympy==1.13.1
terminado==0.18.1
text-unidecode==1.3
threadpoolctl @ file:///home/conda/feedstock_root/build_artifacts/threadpoolctl_1741878222898/work
tifffile @ file:///home/conda/feedstock_root/build_artifacts/tifffile_1734094708987/work
tinycss2==1.4.0
toml @ file:///home/conda/feedstock_root/build_artifacts/toml_1734091811753/work
tomli==2.2.1
toolz @ file:///home/conda/feedstock_root/build_artifacts/toolz_1733736030883/work
torch==2.6.0
torchinfo==1.8.0
torchmetrics==1.7.1
torchvision==0.21.0
tornado @ file:///D:/bld/tornado_1681817664954/work
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
unicodedata2 @ file:///D:/bld/unicodedata2_1667239969015/work
uri-template==1.3.0
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1744323578849/work
uvicorn==0.35.0
watchfiles==1.1.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
win_inet_pton @ file:///D:/bld/win_inet_pton_1733130564612/work
yarl==1.19.0
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1732827521216/work
zstandard==0.19.0
